Offering an alternative way of authenticating with google plus.
Because the google plus api really is our data store from which the
LoginController is pulling, I'm not sure having a /models/Login.js is
necessary. Also, generating models with sails also automatically creates
a mongodb collection by the same name, and we're not storing anything in
it. Although, concievable it could be the place where we could store the
oauth tokens, that's why for now I left the Login model.